+++ modules/datamusic/datamusic-service/src/main/resources/META-INF/sql/tables.sql
@@ -8,7 +8,7 @@ create table datamusic_CollectionItem (
 	userAccountId LONG,
 	mediaId LONG,
 	review VARCHAR(75) null,
-	rating INTEGER
+	rating LONG
 );
 
 create table datamusic_Media (
